Fetch link directly on creation
