--- src/views/links/fetches/show.phtml
@@ -1,69 +1,0 @@
-<?php
-    $this->layout('base.phtml', [
-        'title' => _('Link fetching'),
-        'canonical' => url_full('show fetch link', ['id' => $link->id]),
-    ]);
-?>
-
-<div class="section">
-    <div class="section__title">
-        <h1><?= _('Link fetching') ?></h1>
-    </div>
-
-    <p class="section__intro">
-        <?= protect($link->url) ?>
-    </p>
-
-    <?php if ($link->fetched_at): ?>
-        <p class="paragraph--featured">
-            <?= _f('The link title will be “<em>%s</em>”.', protect($link->title)) ?>
-        </p>
-
-        <p class="paragraph--centered">
-            <a class="anchor--action" href="<?= url('link', ['id' => $link->id]) ?>">
-                <?= _('All is fine, go to the link') ?>
-            </a>
-        </p>
-
-        <p class="paragraph--centered">
-            <a href="<?= url('edit link', ['id' => $link->id]) ?>">
-                <?= _('Let’s change this title') ?>
-            </a>
-        </p>
-    <?php else: ?>
-        <?php if ($error): ?>
-            <p class="form__error">
-                <?= $error ?>
-            </p>
-
-            <form method="post" action="<?= url('fetch link', ['id' => $link->id]) ?>">
-                <input type="hidden" name="csrf" value="<?= csrf_token() ?>" />
-                <button type="submit" class="button--primary">
-                    <?= _('Fetch link information') ?>
-                </button>
-            </form>
-        <?php else: ?>
-            <p class="paragraph--featured">
-                <?= _('Please wait while we’re collecting information…') ?>
-            </p>
-
-            <div class="spinner"></div>
-
-            <form
-                 method="post"
-                 action="<?= url('fetch link', ['id' => $link->id]) ?>"
-                 data-controller="autosubmit"
-                 data-autosubmit-mode="timeout"
-             >
-                <input type="hidden" name="csrf" value="<?= csrf_token() ?>" />
-            </form>
-        <?php endif; ?>
-
-        <img
-            class="illustration illustration--centered"
-            alt=""
-            src="<?= url_static('illustrations/collecting.svg') ?>"
-            height="250"
-        />
-    <?php endif; ?>
-</div>
